Move pool stuff into their own mod
